+++ intersection.c
@@ -1,9 +1,9 @@
 // Checks whether the first rectangle contains one of the two vertices of the second rectangle.
-int PointIntersection(int x1, int x2, int y1, int y2, int x3, int x4, int y3, int y4){
-    int greaterX = (x1 > x2) ? x1 : x2;
-    int greaterY = (y1 > y2) ? y1 : y2;
-    int smallerX = (x1 < x2) ? x1 : x2;
-    int smallerY = (y1 < y2) ? y1 : y2;
+int PointIntersection(long x1, long x2, long y1, long y2, long x3, long x4, long y3, long y4){
+    long greaterX = (x1 > x2) ? x1 : x2;
+    long greaterY = (y1 > y2) ? y1 : y2;
+    long smallerX = (x1 < x2) ? x1 : x2;
+    long smallerY = (y1 < y2) ? y1 : y2;
 
     // Checks for first vertex of the second rectangle...
     // Checks if x3 is in the interval [smallerX, greaterX]
@@ -23,7 +23,7 @@ int PointIntersection(int x1, int x2, int y1, int y2, int x3, int x4, int y3, in
 }
 
 // Checks whether any of the two vertices of one of the rectangles is inside of the other.
-int Intersection(int x1, int y1, int x2, int y2, int x3, int y3, int x4, int y4) {
+int Intersection(long x1, long y1, long x2, long y2, long x3, long y3, long x4, long y4) {
     // Checks if first contains vertex of second or if second contains vertex of first.
     if (PointIntersection(x1,x2,y1,y2,x3,x4,y3,y4) == 1 || PointIntersection(x3,x4,y3,y4,x1,x2,y1,y2) == 1){
         return 1;
