+++ search.c
@@ -1,7 +1,7 @@
 #include <time.h>
 #include "intersection.c"
 
-void searchRectangleInTree(int x1, int y1, int x2, int y2, FILE *file, int initial_fp_pos, int M);
+void searchRectangleInTree(long x1, long y1, long x2, long y2, FILE *file, long initial_fp_pos, int M);
 double searchRectangleFile(char *Q_file_name, char *tree_file_name, int M);
 
 // Searches the entirety of the rectangles in Q_file_name
@@ -12,13 +12,13 @@ double searchRectangleFile(char *Q_file_name, char *tree_file_name, int M) {
     FILE *tree_file;
 
     // Save the contents of the Q_file in an array.
-    int *array_of_ints;
+    long *array_of_ints;
     Q_file = fopen(Q_file_name, "rb");
     fseek(Q_file, 0, SEEK_END);
-    int number_of_ints = ftell(Q_file) / sizeof(int);
+    long number_of_ints = ftell(Q_file) / sizeof(long);
     rewind(Q_file);
-    array_of_ints = (int *) malloc(number_of_ints * sizeof(int));
-    fread(array_of_ints , sizeof(int), number_of_ints, Q_file);
+    array_of_ints = (long *) malloc(number_of_ints * sizeof(long));
+    fread(array_of_ints , sizeof(long), number_of_ints, Q_file);
 
     
     // The performance of the search of all the rectangles in Q is measured.
@@ -28,10 +28,10 @@ double searchRectangleFile(char *Q_file_name, char *tree_file_name, int M) {
     start_time = clock();
 
     for (int i=0; i < number_of_ints; i+=4) {
-        int x1 = array_of_ints[i];
-        int y1 = array_of_ints[i+1];
-        int x2 = array_of_ints[i+2];
-        int y2 = array_of_ints[i+3];
+        long x1 = array_of_ints[i];
+        long y1 = array_of_ints[i+1];
+        long x2 = array_of_ints[i+2];
+        long y2 = array_of_ints[i+3];
         // Search the rectangle in the root of the tree, at position 0 bytes.
         searchRectangleInTree(x1, y1, x2, y2, tree_file, 0, M);
     }
@@ -46,8 +46,8 @@ double searchRectangleFile(char *Q_file_name, char *tree_file_name, int M) {
 
 // Struct that returns the result of a search of a single rectangle in a given tree.
 struct searchResult {
-    int *intersected_rectangles; //array
-    int number_accesses;
+    long *intersected_rectangles; //array
+    long number_accesses;
 };
 
 // Searches a single rectangle in a given tree (contained in the file),
@@ -58,29 +58,29 @@ struct searchResult {
 // representing positions in the file where its children are located.
 // The leaves point to the negative indices of rectangles in the original data R_rect file.
 // Negative numbers in leaves are used to identify leaves in recursion.
-void searchRectangleInTree(int x1, int y1, int x2, int y2, FILE *file, int initial_fp_pos, int M) {
+void searchRectangleInTree(long x1, long y1, long x2, long y2, FILE *file, long initial_fp_pos, int M) {
 //struct searchResult searchRectangleInTree(int x1, int y1, int x2, int y2, FILE *file, int intial_fp_pos, int M) {
 
     // Extract the rectangle's information at the given node.
-    int x1_node;
-    int y1_node;
-    int x2_node;
-    int y2_node;
+    long x1_node;
+    long y1_node;
+    long x2_node;
+    long y2_node;
     fseek(file, initial_fp_pos, SEEK_SET);
-    fread(&x1_node, sizeof(int), 1, file);
-    fread(&y1_node, sizeof(int), 1, file);
-    fread(&x2_node, sizeof(int), 1, file);
-    fread(&y2_node, sizeof(int), 1, file);
+    fread(&x1_node, sizeof(long), 1, file);
+    fread(&y1_node, sizeof(long), 1, file);
+    fread(&x2_node, sizeof(long), 1, file);
+    fread(&y2_node, sizeof(long), 1, file);
 
     // Extract array of file pointers to children
-    int children_fp[M];
+    long children_fp[M];
     for(int k=0; k<M; k++) {
-        fread(&children_fp[k], sizeof(int), 1, file);
+        fread(&children_fp[k], sizeof(long), 1, file);
     }
 
     if (children_fp[0] < 0) { // Leaf
         for(int i =0; i<M; i++) { // Print each index of an intersected rectangle.
-            printf("Intersected rectangle at index %d\n", -children_fp[i]);
+            printf("Intersected rectangle at index %ld\n", -children_fp[i]);
             /*
             children_fp[k] = (-pos_index - 1)*sizeof(int); // Change -index to position in file.
             */
@@ -89,17 +89,17 @@ void searchRectangleInTree(int x1, int y1, int x2, int y2, FILE *file, int initi
 
     else { // Internal node
         for(int i =0; i<M; i++) { // Search in recursion in subtree if need be.
-            int child_fp = children_fp[i];
+            long child_fp = children_fp[i];
             if (child_fp != 0) { // 0 child means no child.
-                int x1_child;
-                int y1_child;
-                int x2_child;
-                int y2_child;
+                long x1_child;
+                long y1_child;
+                long x2_child;
+                long y2_child;
                 fseek(file, child_fp, SEEK_SET);
-                fread(&x1_child, sizeof(int), 1, file);
-                fread(&y1_child, sizeof(int), 1, file);
-                fread(&x2_child, sizeof(int), 1, file);
-                fread(&y2_child, sizeof(int), 1, file);
+                fread(&x1_child, sizeof(long), 1, file);
+                fread(&y1_child, sizeof(long), 1, file);
+                fread(&x2_child, sizeof(long), 1, file);
+                fread(&y2_child, sizeof(long), 1, file);
 
                 int intersection_bool = Intersection(x1, y1, x2, y2, x1_child, y1_child, x2_child, y2_child);
